Add IoProxy and CodecInstanceContainer to allow rust I/O and codec work to begin

+++ imageflow_core/src/flow/nodes/codecs_and_pointer.rs
@@ -112,25 +112,24 @@ fn decoder_def() -> NodeDefinition {
         fn_flatten_pre_optimize: {
             fn f(ctx: &mut OpCtxMut, ix: NodeIndex<u32>) {
 
-//                if ctx.weight(ix).frame_est == FrameEstimate::None{
-//                    decoder_estimate(ctx,ix);
-//                }
-
-                // Mutate instead of replace (custom_state is populated)
+                // Mutate instead of replace
                 ctx.weight_mut(ix).def = &PRIMITIVE_DECODER;
 
+                let io_id = decoder_encoder_io_id(ctx, ix).unwrap();
 
-                let exif_flag = unsafe {ffi::flow_codecs_jpg_decoder_get_exif(ctx.flow_c(), ctx.weight(ix).custom_state as *mut ffi::CodecInstance) };
-                if exif_flag > 0 {
-                    let new_node = ctx.graph
-                        .add_node(Node::new(&APPLY_ORIENTATION,
-                                            NodeParams::Json(s::Node::ApplyOrientation {
-                                                flag: exif_flag,
-                                            })));
-                    ctx.copy_edges_to(ix, new_node, EdgeDirection::Outgoing);
-                    ctx.delete_child_edges_for(ix);
-                    ctx.graph.add_edge(ix, new_node, EdgeKind::Input).unwrap();
+                if let Ok(exif_flag) = ctx.job.get_exif_rotation_flag(io_id){
+                    if exif_flag > 0 {
+                        let new_node = ctx.graph
+                            .add_node(Node::new(&APPLY_ORIENTATION,
+                                                NodeParams::Json(s::Node::ApplyOrientation {
+                                                    flag: exif_flag,
+                                                })));
+                        ctx.copy_edges_to(ix, new_node, EdgeDirection::Outgoing);
+                        ctx.delete_child_edges_for(ix);
+                        ctx.graph.add_edge(ix, new_node, EdgeKind::Input).unwrap();
+                    }
                 }
+
             }
             Some(f)
         },
@@ -147,20 +146,10 @@ fn primitive_decoder_def() -> NodeDefinition {
         fn_link_state_to_this_io_id: Some(decoder_encoder_io_id),
         fn_execute: Some({
             fn f(ctx: &mut OpCtxMut, ix: NodeIndex<u32>) {
+                let io_id = decoder_encoder_io_id(ctx, ix).unwrap();
 
-                // TODO______
-                let codec = ctx.weight(ix).custom_state as *mut ffi::CodecInstance;
-                if codec.is_null(){
-                    panic!("custom_state: CodecInstance is null. Codec failed to link?")
-                }
-                unsafe {
-                    let result = ffi::flow_codec_execute_read_frame(ctx.flow_c(), codec);
-                    if result.is_null() {
-                        ctx.assert_ok();
-                    } else {
-                        ctx.weight_mut(ix).result = NodeResult::Frame(result);
-                    }
-                }
+                let result = ctx.job.get_codec(io_id).unwrap().read_frame(ctx.c, ctx.job).unwrap();
+                ctx.weight_mut(ix).result = NodeResult::Frame(result);
             }
             f
         }),
@@ -180,76 +169,19 @@ fn encoder_def() -> NodeDefinition {
 
         fn_execute: Some({
             fn f(ctx: &mut OpCtxMut, ix: NodeIndex<u32>) {
+                let io_id = decoder_encoder_io_id(ctx, ix).unwrap();
+
                 if let Some(input_bitmap) = ctx.first_parent_result_frame(ix, EdgeKind::Input) {
                     let result;
                     {
                         let weight = ctx.weight(ix);
                         if let NodeParams::Json(s::Node::Encode { ref preset, ref io_id, .. }) =
                                weight.params {
-                            let codec = weight.custom_state as *mut ffi::CodecInstance;
 
-                            if codec.is_null(){
-                                panic!("custom_state: CodecInstance is null")
-                            }
+                            result = NodeResult::Encoded(
+                                ctx.job.get_codec(*io_id).unwrap().write_frame(ctx.c, ctx.job, preset,
+                                                  unsafe{ &mut *input_bitmap } ).unwrap());
 
-                            let (wanted_id, hints) = match *preset {
-                                s::EncoderPreset::LibjpegTurbo { quality } => {
-                                    (ffi::CodecType::EncodeJpeg as i64,
-                                     ffi::EncoderHints {
-                                        jpeg_encode_quality: quality.unwrap_or(90),
-                                        disable_png_alpha: false,
-                                    })
-                                }
-                                s::EncoderPreset::Libpng { ref matte,
-                                                           zlib_compression,
-                                                           ref depth } => {
-                                    (ffi::CodecType::EncodePng as i64,
-                                     ffi::EncoderHints {
-                                        jpeg_encode_quality: -1,
-                                        disable_png_alpha: match *depth {
-                                            Some(s::PngBitDepth::Png24) => true,
-                                            _ => false,
-                                        },
-                                    })
-                                }
-                            };
-
-                            unsafe {
-                                let (result_mime, result_ext) = match *preset {
-                                    s::EncoderPreset::Libpng { .. } => ("image/png", "png"),
-                                    s::EncoderPreset::LibjpegTurbo { .. } => ("image/jpeg", "jpg"),
-                                };
-                                result = NodeResult::Encoded(s::EncodeResult {
-                                    w: (*input_bitmap).w as i32,
-                                    h: (*input_bitmap).h as i32,
-                                    preferred_mime_type: result_mime.to_owned(),
-                                    preferred_extension: result_ext.to_owned(),
-                                    io_id: *io_id,
-                                    bytes: s::ResultBytes::Elsewhere,
-                                });
-
-
-
-                                (*codec).codec_id = wanted_id;
-                                if !ffi::flow_codec_initialize(ctx.flow_c(), codec) {
-                                    ctx.assert_ok();
-                                }
-                                let codec_def = ffi::flow_codec_get_definition(ctx.flow_c(), wanted_id);
-                                if codec_def.is_null() {
-                                    ctx.assert_ok();
-                                }
-                                let write_fn = (*codec_def).write_frame;
-                                if write_fn == None {
-                                    panic!("Codec didn't implement write_frame");
-                                }
-
-                                if !write_fn.unwrap()(ctx.flow_c(),
-                                                      (*codec).codec_state,
-                                                      input_bitmap,
-                                                      &hints as *const ffi::EncoderHints) {
-                                    ctx.assert_ok();
-                                }
-                            }
                         } else {
                             panic!("");
                         }
